+++ es-serverless-20230627/alibabacloud_es_serverless20230627/models.py
@@ -306,12 +306,17 @@ class CreateAppRequest(TeaModel):
         private_network: List[CreateAppRequestPrivateNetwork] = None,
         quota_info: CreateAppRequestQuotaInfo = None,
         region_id: str = None,
+        scenario: str = None,
         version: str = None,
         dry_run: bool = None,
     ):
         # 应用名
+        # 
+        # This parameter is required.
         self.app_name = app_name
+        # This parameter is required.
         self.authentication = authentication
+        # This parameter is required.
         self.charge_type = charge_type
         # 应用备注
         self.description = description
@@ -319,6 +324,7 @@ class CreateAppRequest(TeaModel):
         self.private_network = private_network
         self.quota_info = quota_info
         self.region_id = region_id
+        self.scenario = scenario
         self.version = version
         self.dry_run = dry_run
 
@@ -362,6 +368,8 @@ class CreateAppRequest(TeaModel):
             result['quotaInfo'] = self.quota_info.to_map()
         if self.region_id is not None:
             result['regionId'] = self.region_id
+        if self.scenario is not None:
+            result['scenario'] = self.scenario
         if self.version is not None:
             result['version'] = self.version
         if self.dry_run is not None:
@@ -394,6 +402,8 @@ class CreateAppRequest(TeaModel):
             self.quota_info = temp_model.from_map(m['quotaInfo'])
         if m.get('regionId') is not None:
             self.region_id = m.get('regionId')
+        if m.get('scenario') is not None:
+            self.scenario = m.get('scenario')
         if m.get('version') is not None:
             self.version = m.get('version')
         if m.get('dryRun') is not None:
@@ -475,9 +485,6 @@ class CreateAppResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -503,6 +510,198 @@ class CreateAppResponse(TeaModel):
             self.status_code = m.get('statusCode')
         if m.get('body') is not None:
             temp_model = CreateAppResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
+class CreateEndpointRequestEndpointZones(TeaModel):
+    def __init__(
+        self,
+        vswitch_id: str = None,
+        zone_id: str = None,
+    ):
+        self.vswitch_id = vswitch_id
+        self.zone_id = zone_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.vswitch_id is not None:
+            result['vswitchId'] = self.vswitch_id
+        if self.zone_id is not None:
+            result['zoneId'] = self.zone_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('vswitchId') is not None:
+            self.vswitch_id = m.get('vswitchId')
+        if m.get('zoneId') is not None:
+            self.zone_id = m.get('zoneId')
+        return self
+
+
+class CreateEndpointRequest(TeaModel):
+    def __init__(
+        self,
+        endpoint_zones: List[CreateEndpointRequestEndpointZones] = None,
+        name: str = None,
+        vpc_id: str = None,
+        type: str = None,
+    ):
+        # This parameter is required.
+        self.endpoint_zones = endpoint_zones
+        self.name = name
+        # This parameter is required.
+        self.vpc_id = vpc_id
+        self.type = type
+
+    def validate(self):
+        if self.endpoint_zones:
+            for k in self.endpoint_zones:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        result['endpointZones'] = []
+        if self.endpoint_zones is not None:
+            for k in self.endpoint_zones:
+                result['endpointZones'].append(k.to_map() if k else None)
+        if self.name is not None:
+            result['name'] = self.name
+        if self.vpc_id is not None:
+            result['vpcId'] = self.vpc_id
+        if self.type is not None:
+            result['type'] = self.type
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        self.endpoint_zones = []
+        if m.get('endpointZones') is not None:
+            for k in m.get('endpointZones'):
+                temp_model = CreateEndpointRequestEndpointZones()
+                self.endpoint_zones.append(temp_model.from_map(k))
+        if m.get('name') is not None:
+            self.name = m.get('name')
+        if m.get('vpcId') is not None:
+            self.vpc_id = m.get('vpcId')
+        if m.get('type') is not None:
+            self.type = m.get('type')
+        return self
+
+
+class CreateEndpointResponseBodyResult(TeaModel):
+    def __init__(
+        self,
+        endpoint_id: str = None,
+    ):
+        self.endpoint_id = endpoint_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.endpoint_id is not None:
+            result['endpointId'] = self.endpoint_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('endpointId') is not None:
+            self.endpoint_id = m.get('endpointId')
+        return self
+
+
+class CreateEndpointResponseBody(TeaModel):
+    def __init__(
+        self,
+        request_id: str = None,
+        result: CreateEndpointResponseBodyResult = None,
+    ):
+        # Id of the request
+        self.request_id = request_id
+        self.result = result
+
+    def validate(self):
+        if self.result:
+            self.result.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.request_id is not None:
+            result['requestId'] = self.request_id
+        if self.result is not None:
+            result['result'] = self.result.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('requestId') is not None:
+            self.request_id = m.get('requestId')
+        if m.get('result') is not None:
+            temp_model = CreateEndpointResponseBodyResult()
+            self.result = temp_model.from_map(m['result'])
+        return self
+
+
+class CreateEndpointResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: CreateEndpointResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = CreateEndpointResponseBody()
             self.body = temp_model.from_map(m['body'])
         return self
 
@@ -581,9 +780,6 @@ class DeleteAppResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -640,6 +836,202 @@ class GetAppRequest(TeaModel):
         return self
 
 
+class GetAppResponseBodyResultNetworkWhiteIpGroup(TeaModel):
+    def __init__(
+        self,
+        group_name: str = None,
+        ips: List[str] = None,
+    ):
+        self.group_name = group_name
+        self.ips = ips
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.group_name is not None:
+            result['groupName'] = self.group_name
+        if self.ips is not None:
+            result['ips'] = self.ips
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('groupName') is not None:
+            self.group_name = m.get('groupName')
+        if m.get('ips') is not None:
+            self.ips = m.get('ips')
+        return self
+
+
+class GetAppResponseBodyResultNetwork(TeaModel):
+    def __init__(
+        self,
+        domain: str = None,
+        enabled: bool = None,
+        port: int = None,
+        type: str = None,
+        white_ip_group: List[GetAppResponseBodyResultNetworkWhiteIpGroup] = None,
+    ):
+        self.domain = domain
+        self.enabled = enabled
+        self.port = port
+        self.type = type
+        self.white_ip_group = white_ip_group
+
+    def validate(self):
+        if self.white_ip_group:
+            for k in self.white_ip_group:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.domain is not None:
+            result['domain'] = self.domain
+        if self.enabled is not None:
+            result['enabled'] = self.enabled
+        if self.port is not None:
+            result['port'] = self.port
+        if self.type is not None:
+            result['type'] = self.type
+        result['whiteIpGroup'] = []
+        if self.white_ip_group is not None:
+            for k in self.white_ip_group:
+                result['whiteIpGroup'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('domain') is not None:
+            self.domain = m.get('domain')
+        if m.get('enabled') is not None:
+            self.enabled = m.get('enabled')
+        if m.get('port') is not None:
+            self.port = m.get('port')
+        if m.get('type') is not None:
+            self.type = m.get('type')
+        self.white_ip_group = []
+        if m.get('whiteIpGroup') is not None:
+            for k in m.get('whiteIpGroup'):
+                temp_model = GetAppResponseBodyResultNetworkWhiteIpGroup()
+                self.white_ip_group.append(temp_model.from_map(k))
+        return self
+
+
+class GetAppResponseBodyResultPrivateNetworkWhiteIpGroup(TeaModel):
+    def __init__(
+        self,
+        group_name: str = None,
+        ips: List[str] = None,
+    ):
+        self.group_name = group_name
+        self.ips = ips
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.group_name is not None:
+            result['groupName'] = self.group_name
+        if self.ips is not None:
+            result['ips'] = self.ips
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('groupName') is not None:
+            self.group_name = m.get('groupName')
+        if m.get('ips') is not None:
+            self.ips = m.get('ips')
+        return self
+
+
+class GetAppResponseBodyResultPrivateNetwork(TeaModel):
+    def __init__(
+        self,
+        domain: str = None,
+        enabled: bool = None,
+        port: int = None,
+        pvl_endpoint_id: str = None,
+        type: str = None,
+        vpc_id: str = None,
+        white_ip_group: List[GetAppResponseBodyResultPrivateNetworkWhiteIpGroup] = None,
+    ):
+        self.domain = domain
+        self.enabled = enabled
+        self.port = port
+        self.pvl_endpoint_id = pvl_endpoint_id
+        self.type = type
+        self.vpc_id = vpc_id
+        self.white_ip_group = white_ip_group
+
+    def validate(self):
+        if self.white_ip_group:
+            for k in self.white_ip_group:
+                if k:
+                    k.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.domain is not None:
+            result['domain'] = self.domain
+        if self.enabled is not None:
+            result['enabled'] = self.enabled
+        if self.port is not None:
+            result['port'] = self.port
+        if self.pvl_endpoint_id is not None:
+            result['pvlEndpointId'] = self.pvl_endpoint_id
+        if self.type is not None:
+            result['type'] = self.type
+        if self.vpc_id is not None:
+            result['vpcId'] = self.vpc_id
+        result['whiteIpGroup'] = []
+        if self.white_ip_group is not None:
+            for k in self.white_ip_group:
+                result['whiteIpGroup'].append(k.to_map() if k else None)
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('domain') is not None:
+            self.domain = m.get('domain')
+        if m.get('enabled') is not None:
+            self.enabled = m.get('enabled')
+        if m.get('port') is not None:
+            self.port = m.get('port')
+        if m.get('pvlEndpointId') is not None:
+            self.pvl_endpoint_id = m.get('pvlEndpointId')
+        if m.get('type') is not None:
+            self.type = m.get('type')
+        if m.get('vpcId') is not None:
+            self.vpc_id = m.get('vpcId')
+        self.white_ip_group = []
+        if m.get('whiteIpGroup') is not None:
+            for k in m.get('whiteIpGroup'):
+                temp_model = GetAppResponseBodyResultPrivateNetworkWhiteIpGroup()
+                self.white_ip_group.append(temp_model.from_map(k))
+        return self
+
+
 class GetAppResponseBodyResult(TeaModel):
     def __init__(
         self,
@@ -649,7 +1041,9 @@ class GetAppResponseBodyResult(TeaModel):
         description: str = None,
         instance_id: str = None,
         modified_time: str = None,
+        network: List[GetAppResponseBodyResultNetwork] = None,
         owner_id: str = None,
+        private_network: List[GetAppResponseBodyResultPrivateNetwork] = None,
         region_id: str = None,
         status: str = None,
         version: str = None,
@@ -660,13 +1054,22 @@ class GetAppResponseBodyResult(TeaModel):
         self.description = description
         self.instance_id = instance_id
         self.modified_time = modified_time
+        self.network = network
         self.owner_id = owner_id
+        self.private_network = private_network
         self.region_id = region_id
         self.status = status
         self.version = version
 
     def validate(self):
-        pass
+        if self.network:
+            for k in self.network:
+                if k:
+                    k.validate()
+        if self.private_network:
+            for k in self.private_network:
+                if k:
+                    k.validate()
 
     def to_map(self):
         _map = super().to_map()
@@ -686,8 +1089,16 @@ class GetAppResponseBodyResult(TeaModel):
             result['instanceId'] = self.instance_id
         if self.modified_time is not None:
             result['modifiedTime'] = self.modified_time
+        result['network'] = []
+        if self.network is not None:
+            for k in self.network:
+                result['network'].append(k.to_map() if k else None)
         if self.owner_id is not None:
             result['ownerId'] = self.owner_id
+        result['privateNetwork'] = []
+        if self.private_network is not None:
+            for k in self.private_network:
+                result['privateNetwork'].append(k.to_map() if k else None)
         if self.region_id is not None:
             result['regionId'] = self.region_id
         if self.status is not None:
@@ -710,8 +1121,18 @@ class GetAppResponseBodyResult(TeaModel):
             self.instance_id = m.get('instanceId')
         if m.get('modifiedTime') is not None:
             self.modified_time = m.get('modifiedTime')
+        self.network = []
+        if m.get('network') is not None:
+            for k in m.get('network'):
+                temp_model = GetAppResponseBodyResultNetwork()
+                self.network.append(temp_model.from_map(k))
         if m.get('ownerId') is not None:
             self.owner_id = m.get('ownerId')
+        self.private_network = []
+        if m.get('privateNetwork') is not None:
+            for k in m.get('privateNetwork'):
+                temp_model = GetAppResponseBodyResultPrivateNetwork()
+                self.private_network.append(temp_model.from_map(k))
         if m.get('regionId') is not None:
             self.region_id = m.get('regionId')
         if m.get('status') is not None:
@@ -768,9 +1189,6 @@ class GetAppResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -962,9 +1380,6 @@ class GetAppQuotaResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1149,9 +1564,6 @@ class GetMonitorDataResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1340,6 +1752,7 @@ class ListAppsResponseBody(TeaModel):
     ):
         self.request_id = request_id
         self.result = result
+        # This parameter is required.
         self.total_count = total_count
 
     def validate(self):
@@ -1390,9 +1803,6 @@ class ListAppsResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()
 
@@ -1915,9 +2325,6 @@ class UpdateAppResponse(TeaModel):
         self.body = body
 
     def validate(self):
-        self.validate_required(self.headers, 'headers')
-        self.validate_required(self.status_code, 'status_code')
-        self.validate_required(self.body, 'body')
         if self.body:
             self.body.validate()