+++ es-serverless-20230627/ChangeLog.md
@@ -1,3 +1,12 @@
+2024-05-23 Version: 2.0.0
+- Support API CreateEndpoint.
+- Update API CreateApp: add param body.
+- Update API CreateApp: delete param Body.
+- Update API GetApp: update response param.
+- Update API UpdateApp: add param body.
+- Update API UpdateApp: delete param Body.
+
+
 2024-01-24 Version: 1.0.2
 - Generated python 2023-06-27 for es-serverless.